feat: TransactionOverview component [BLOX-205]
